feature: Added new features to sample app
