+++ example/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.Example"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".presentation.home.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Example">
