+++ example/src/main/java/com/mercadopago/sdk/android/example/data/model/Installment.kt
@@ -1,0 +1,5 @@
+package com.mercadopago.sdk.android.example.data.model
+
+internal data class Installment(
+    val value: String
+)
