+++ example/src/main/java/com/mercadopago/sdk/android/example/domain/model/LogEntry.kt
@@ -1,0 +1,8 @@
+package com.mercadopago.sdk.android.example.domain.model
+
+data class LogEntry(
+    val type: LogType,
+    val message: String,
+    val response: String,
+    val timestamp: String,
+)
