+++ example/src/main/java/com/mercadopago/sdk/android/example/domain/model/LogType.kt
@@ -1,0 +1,7 @@
+package com.mercadopago.sdk.android.example.domain.model
+
+enum class LogType {
+    All,
+    Network,
+    Function,
+}
