+++ example/src/main/java/com/mercadopago/sdk/android/example/domain/model/PublicKey.kt
@@ -1,0 +1,9 @@
+package com.mercadopago.sdk.android.example.domain.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+internal data class PublicKey(
+    val publicKey: String,
+    val countryCode: String,
+)
