+++ example/src/main/java/com/mercadopago/sdk/android/example/presentation/coremethods/PaymentExampleScreen.kt
@@ -1,0 +1,431 @@
+package com.mercadopago.sdk.android.example.presentation.coremethods
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.text.selection.LocalTextSelectionColors
+import androidx.compose.foundation.text.selection.TextSelectionColors
+import androidx.compose.material3.BottomSheetScaffold
+import androidx.compose.material3.Button
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.OutlinedTextFieldDefaults
+import androidx.compose.material3.Scaffold
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.SolidColor
+import androidx.compose.ui.platform.LocalClipboardManager
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.AnnotatedString
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import coil3.compose.AsyncImage
+import coil3.request.ImageRequest
+import com.mercadopago.sdk.android.coremethods.domain.model.IdentificationType
+import com.mercadopago.sdk.android.coremethods.ui.components.textfield.cardnumber.CardNumberTextField
+import com.mercadopago.sdk.android.coremethods.ui.components.textfield.cardnumber.CardNumberTextFieldEvent
+import com.mercadopago.sdk.android.coremethods.ui.components.textfield.expirationdate.ExpirationDateFormat
+import com.mercadopago.sdk.android.coremethods.ui.components.textfield.expirationdate.ExpirationDateTextField
+import com.mercadopago.sdk.android.coremethods.ui.components.textfield.expirationdate.ExpirationDateTextFieldEvent
+import com.mercadopago.sdk.android.coremethods.ui.components.textfield.pcitextfield.PCIFieldState
+import com.mercadopago.sdk.android.coremethods.ui.components.textfield.pcitextfield.rememberPCIFieldState
+import com.mercadopago.sdk.android.coremethods.ui.components.textfield.securitycode.SecurityCodeTextField
+import com.mercadopago.sdk.android.coremethods.ui.components.textfield.securitycode.SecurityCodeTextFieldEvent
+import com.mercadopago.sdk.android.coremethods.ui.utils.MaskVisualTransformation
+import com.mercadopago.sdk.android.example.R
+import com.mercadopago.sdk.android.example.extensions.addBorder
+import com.mercadopago.sdk.android.example.data.model.Installment
+import com.mercadopago.sdk.android.example.presentation.components.AlertDialog
+import com.mercadopago.sdk.android.example.presentation.coremethods.state.CardNumberTextFieldState
+import com.mercadopago.sdk.android.example.presentation.coremethods.state.ExpirationDateState
+import com.mercadopago.sdk.android.example.presentation.coremethods.state.IdentificationState
+import com.mercadopago.sdk.android.example.presentation.coremethods.state.PaymentScreenDialogState
+import com.mercadopago.sdk.android.example.presentation.coremethods.state.PaymentScreenViewState
+import com.mercadopago.sdk.android.example.presentation.coremethods.state.SecurityCodeState
+import com.mercadopago.sdk.android.example.presentation.components.CardTokenDialog
+import com.mercadopago.sdk.android.example.presentation.components.IdentificationTypeSelectorField
+import com.mercadopago.sdk.android.example.presentation.components.InstallmentListDropDownField
+import com.mercadopago.sdk.android.example.presentation.components.Label
+import com.mercadopago.sdk.android.example.presentation.components.PlaceHolder
+import com.mercadopago.sdk.android.example.presentation.logs.DebugLogsScreen
+import com.mercadopago.sdk.android.example.presentation.theme.MercadoPagoSampleTheme
+import org.koin.androidx.compose.koinViewModel
+
+@Composable
+internal fun PaymentExampleScreen(
+    viewModel: PaymentScreenViewModel = koinViewModel(),
+) {
+    val viewState by viewModel.viewState.collectAsState()
+    val cardNumberState = rememberPCIFieldState()
+    val expirationDateState = rememberPCIFieldState()
+    val securityCodeState = rememberPCIFieldState()
+
+    LaunchedEffect(key1 = true) {
+        viewModel.getIdentificationTypes()
+    }
+
+    PaymentExampleScreenContent(
+        viewState = viewState,
+        cardNumberState = cardNumberState,
+        expirationDateState = expirationDateState,
+        securityCodeState = securityCodeState,
+        onGenerateCardToken = {
+            viewModel.generateToken(
+                cardNumberState = cardNumberState,
+                expirationDateState = expirationDateState,
+                securityCodeState = securityCodeState,
+            )
+        },
+        onExpirationDateEvent = viewModel::onExpirationDateEvent,
+        onSecurityCodeEvent = viewModel::onSecurityCodeEvent,
+        onCardNumberEvent = viewModel::onCardNumberEvent,
+        onSelectIdentification = viewModel::onIdentificationTypeChanged,
+        onIdentificationTypeChanged = viewModel::onIdentificationTypeValueChanged,
+        onSelectedInstallment = viewModel::onInstallmentSelected,
+        onCardHolderNameChanged = viewModel::onCardHolderNameChanged,
+        onDialogStateChange = viewModel::onDialogStateChanged
+    )
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+internal fun PaymentExampleScreenContent(
+    modifier: Modifier = Modifier,
+    viewState: PaymentScreenViewState,
+    cardNumberState: PCIFieldState,
+    expirationDateState: PCIFieldState,
+    securityCodeState: PCIFieldState,
+    onGenerateCardToken: () -> Unit,
+    onExpirationDateEvent: (ExpirationDateTextFieldEvent) -> Unit,
+    onSecurityCodeEvent: (SecurityCodeTextFieldEvent) -> Unit,
+    onCardNumberEvent: (CardNumberTextFieldEvent) -> Unit,
+    onSelectIdentification: (IdentificationType) -> Unit,
+    onIdentificationTypeChanged: (String) -> Unit,
+    onSelectedInstallment: (Installment) -> Unit,
+    onCardHolderNameChanged: (String) -> Unit,
+    onDialogStateChange: (PaymentScreenDialogState) -> Unit,
+) {
+    val clipboardManager = LocalClipboardManager.current
+    when (viewState.dialogState) {
+        is PaymentScreenDialogState.CardToken -> CardTokenDialog(
+            token = viewState.dialogState.token,
+            onDismiss = {
+                onDialogStateChange(PaymentScreenDialogState.Hidden)
+                clipboardManager.setText(AnnotatedString(viewState.dialogState.token))
+            },
+        )
+        is PaymentScreenDialogState.Error -> AlertDialog(
+            title = viewState.dialogState.title,
+            description = viewState.dialogState.description,
+            onDismissRequest = {
+                onDialogStateChange(PaymentScreenDialogState.Hidden)
+            },
+        )
+
+        PaymentScreenDialogState.Hidden -> Unit
+    }
+
+    BottomSheetScaffold(
+        sheetContent = {
+            DebugLogsScreen()
+        }
+    ) {
+        Scaffold(modifier = modifier.fillMaxSize()) { paddingValues ->
+            Box(
+                Modifier
+                    .fillMaxSize()
+                    .padding(paddingValues)
+            ) {
+                Column {
+                    Spacer(Modifier.height(16.dp))
+                    CardNumberTextFieldExample(
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .padding(horizontal = 16.dp),
+                        state = cardNumberState,
+                        cardNumberState = viewState.cardNumberState,
+                        onCardNumberEvent = onCardNumberEvent
+                    )
+                    Spacer(Modifier.height(16.dp))
+                    Row(modifier = Modifier.padding(horizontal = 16.dp)) {
+                        ExpirationDateExample(
+                            Modifier.weight(1f),
+                            state = expirationDateState,
+                            expirationDateState = viewState.expirationDateState,
+                            onExpirationDateEvent = onExpirationDateEvent
+                        )
+                        Spacer(Modifier.width(16.dp))
+                        SecurityCodeExample(
+                            Modifier.weight(1f),
+                            state = securityCodeState,
+                            securityCodeState = viewState.secureCodeState,
+                            onSecurityCodeEvent = onSecurityCodeEvent
+                        )
+                    }
+                    IdentificationName(
+                        modifier = Modifier.padding(horizontal = 16.dp),
+                        identificationState = viewState.identificationState,
+                        onCardHolderNameChanged = onCardHolderNameChanged
+                    )
+                    Spacer(Modifier.size(16.dp))
+                    IdentificationTypeSelectorField(
+                        state = viewState.identificationState,
+                        onSelectIdentification = onSelectIdentification,
+                        onIdentificationTypeChanged = onIdentificationTypeChanged
+                    )
+                    Spacer(Modifier.size(16.dp))
+                    InstallmentListDropDownField(
+                        state = viewState.installmentsState,
+                        onSelectedInstallment = onSelectedInstallment
+                    )
+                    Spacer(Modifier.size(16.dp))
+
+                    Button(
+                        shape = MaterialTheme.shapes.small,
+                        onClick = onGenerateCardToken,
+                        modifier = Modifier
+                            .align(Alignment.End)
+                            .fillMaxWidth()
+                            .padding(horizontal = 16.dp),
+                    ) {
+                        Row(
+                            verticalAlignment = Alignment.CenterVertically,
+                        ) {
+                            Icon(
+                                painter = painterResource(R.drawable.ic_padlock_closed),
+                                tint = MaterialTheme.colorScheme.onPrimary,
+                                contentDescription = null,
+                            )
+                            Spacer(Modifier.width(4.dp))
+                            Label(
+                                text = "Pay",
+                                textColor = MaterialTheme.colorScheme.onPrimary,
+                                textStyle = MaterialTheme.typography.bodyMedium.copy(fontWeight = FontWeight.Bold),
+                            )
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
+
+@Composable
+internal fun SecurityCodeExample(
+    modifier: Modifier = Modifier,
+    state: PCIFieldState,
+    securityCodeState: SecurityCodeState,
+    onSecurityCodeEvent: (SecurityCodeTextFieldEvent) -> Unit
+) {
+    Column(modifier = modifier) {
+        Label(
+            text = "Security code",
+        )
+        Spacer(Modifier.height(4.dp))
+        CompositionLocalProvider(
+            LocalTextSelectionColors provides TextSelectionColors(
+                MaterialTheme.colorScheme.primary,
+                MaterialTheme.colorScheme.tertiaryContainer
+            )
+        ) {
+            SecurityCodeTextField(
+                state = state,
+                onEvent = onSecurityCodeEvent,
+                textStyle = MaterialTheme.typography.bodyMedium.copy(
+                    color = MaterialTheme.colorScheme.onBackground,
+                ),
+                securityCodeSize = securityCodeState.secureCodeLength,
+                cursorBrush = SolidColor(MaterialTheme.colorScheme.primary),
+                decorationBox = { innerTextField ->
+                    Row(
+                        verticalAlignment = Alignment.CenterVertically,
+                        modifier = Modifier
+                            .addBorder(
+                                isFocused = securityCodeState.isFocused,
+                            )
+                            .height(OutlinedTextFieldDefaults.MinHeight)
+                            .padding(horizontal = 16.dp),
+                    ) {
+                        Box(modifier = Modifier.weight(1f)) {
+                            if (securityCodeState.length == 0) {
+                                PlaceHolder(
+                                    text = "123",
+                                    modifier = Modifier.align(Alignment.CenterStart),
+                                )
+                            }
+                            innerTextField()
+                        }
+                        Spacer(Modifier.width(4.dp))
+                        Image(
+                            painter = painterResource(R.drawable.ic_security_code),
+                            contentDescription = null,
+                            modifier = Modifier.size(34.dp),
+                        )
+                    }
+                }
+            )
+        }
+    }
+}
+
+@Composable
+internal fun ExpirationDateExample(
+    modifier: Modifier = Modifier,
+    state: PCIFieldState,
+    expirationDateState: ExpirationDateState,
+    onExpirationDateEvent: (ExpirationDateTextFieldEvent) -> Unit
+) {
+    Column(modifier = modifier) {
+        Label(
+            text = "Expiration Date",
+        )
+        Spacer(Modifier.height(4.dp))
+        CompositionLocalProvider(
+            LocalTextSelectionColors provides TextSelectionColors(
+                MaterialTheme.colorScheme.primary,
+                MaterialTheme.colorScheme.tertiaryContainer
+            )
+        ) {
+            ExpirationDateTextField(
+                modifier = Modifier.fillMaxWidth(),
+                state = state,
+                dateFormat = ExpirationDateFormat.ShortFormat,
+                onEvent = onExpirationDateEvent,
+                textStyle = MaterialTheme.typography.bodyMedium.copy(
+                    color = MaterialTheme.colorScheme.onBackground,
+                ),
+                decorationBox = { innerTextField ->
+                    Row(
+                        verticalAlignment = Alignment.CenterVertically,
+                        modifier = Modifier
+                            .addBorder(
+                                isFocused = expirationDateState.isFocused,
+                                isValid = expirationDateState.valid
+                            )
+                            .height(OutlinedTextFieldDefaults.MinHeight)
+                            .padding(horizontal = 16.dp),
+                    ) {
+                        Box {
+                            if (expirationDateState.length == 0) {
+                                PlaceHolder(
+                                    text = "MM/YY",
+                                    modifier = Modifier.align(Alignment.CenterStart),
+                                )
+                            }
+                            innerTextField()
+                        }
+                    }
+                }
+            )
+        }
+    }
+}
+
+@Composable
+internal fun CardNumberTextFieldExample(
+    modifier: Modifier = Modifier,
+    state: PCIFieldState,
+    cardNumberState: CardNumberTextFieldState,
+    onCardNumberEvent: (CardNumberTextFieldEvent) -> Unit
+) {
+    Column(modifier = modifier) {
+        Label(
+            text = "Card Number",
+        )
+        Spacer(Modifier.height(4.dp))
+        CardNumberTextField(
+            state = state,
+            onEvent = onCardNumberEvent,
+            textStyle = MaterialTheme.typography.bodyMedium.copy(
+                color = MaterialTheme.colorScheme.onBackground,
+            ),
+            maxLength = cardNumberState.maxLength,
+            visualTransformation = MaskVisualTransformation(cardNumberState.mask),
+            decorationBox = { innerTextField ->
+                Row(
+                    verticalAlignment = Alignment.CenterVertically,
+                    modifier = Modifier
+                        .addBorder(
+                            isFocused = cardNumberState.isFocused,
+                            isValid = cardNumberState.isError.not(),
+                        )
+                        .height(OutlinedTextFieldDefaults.MinHeight)
+                        .padding(horizontal = 16.dp),
+                ) {
+                    // Displaying the issuer image to the user when the bin is changed
+                    if (cardNumberState.image != null) {
+                        val context = LocalContext.current
+                        AsyncImage(
+                            model = ImageRequest.Builder(context)
+                                .data(cardNumberState.image)
+                                .build(),
+                            contentDescription = null,
+                            modifier = Modifier.size(32.dp)
+                        )
+                        Spacer(Modifier.width(4.dp))
+                    }
+                    Box {
+                        if (cardNumberState.length == 0) {
+                            PlaceHolder(
+                                text = "4444 4444 4444 4444",
+                                modifier = Modifier.align(Alignment.CenterStart),
+                            )
+                        }
+                        innerTextField()
+                    }
+                }
+            },
+            modifier = Modifier.fillMaxWidth(),
+        )
+    }
+}
+
+@Composable
+internal fun IdentificationName(
+    modifier: Modifier = Modifier,
+    identificationState: IdentificationState,
+    onCardHolderNameChanged: (String) -> Unit
+) {
+    Column(modifier = modifier) {
+        Spacer(Modifier.height(8.dp))
+        Label(text = "Cardholder's name as it appears on the card")
+        Spacer(Modifier.height(8.dp))
+        OutlinedTextField(
+            value = identificationState.identificationNameValue,
+            placeholder = {
+                PlaceHolder(text = "María López")
+            },
+            textStyle = MaterialTheme.typography.bodyMedium.copy(
+                color = MaterialTheme.colorScheme.onBackground,
+            ),
+            onValueChange = onCardHolderNameChanged,
+            shape = MaterialTheme.shapes.small,
+            modifier = Modifier.fillMaxWidth(),
+        )
+    }
+}
+
+@Preview(name = "Payment Screen Example", showBackground = true)
+@Composable
+internal fun PaymentExampleScreenPreview() {
+    MercadoPagoSampleTheme {
+        PaymentExampleScreen()
+    }
+}
