+++ example/src/main/java/com/mercadopago/sdk/android/example/presentation/coremethods/PaymentScreenViewModel.kt
@@ -1,0 +1,486 @@
+package com.mercadopago.sdk.android.example.presentation.coremethods
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.google.gson.GsonBuilder
+import com.mercadopago.sdk.android.coremethods.domain.interactor.CoreMethods
+import com.mercadopago.sdk.android.coremethods.domain.interactor.coreMethods
+import com.mercadopago.sdk.android.coremethods.domain.model.BuyerIdentification
+import com.mercadopago.sdk.android.coremethods.domain.model.IdentificationType
+import com.mercadopago.sdk.android.coremethods.domain.model.ResultError
+import com.mercadopago.sdk.android.coremethods.domain.utils.Result
+import com.mercadopago.sdk.android.coremethods.ui.components.textfield.cardnumber.CardNumberTextFieldEvent
+import com.mercadopago.sdk.android.coremethods.ui.components.textfield.expirationdate.ExpirationDateTextFieldEvent
+import com.mercadopago.sdk.android.coremethods.ui.components.textfield.pcitextfield.PCIFieldState
+import com.mercadopago.sdk.android.coremethods.ui.components.textfield.securitycode.SecurityCodeTextFieldEvent
+import com.mercadopago.sdk.android.example.data.mappers.toInstallmentModel
+import com.mercadopago.sdk.android.example.data.model.Installment
+import com.mercadopago.sdk.android.example.domain.model.LogType
+import com.mercadopago.sdk.android.example.presentation.coremethods.state.PaymentScreenDialogState
+import com.mercadopago.sdk.android.example.presentation.coremethods.state.PaymentScreenViewState
+import com.mercadopago.sdk.android.example.utils.LogHelper
+import com.mercadopago.sdk.android.initializer.MercadoPagoSDK
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.launch
+import java.math.BigDecimal
+
+internal const val CARD_NUMBER_BIN_LENGTH = 6
+internal const val DEFAULT_MAX_CARD_LENGTH = 19
+
+internal class PaymentScreenViewModel(
+    private val coreMethods: CoreMethods = MercadoPagoSDK.getInstance().coreMethods,
+) : ViewModel() {
+
+    private val _viewState = MutableStateFlow(PaymentScreenViewState())
+    val viewState: StateFlow<PaymentScreenViewState> = _viewState
+
+    private val gson = GsonBuilder()
+        .setPrettyPrinting()
+        .create()
+
+    fun generateToken(
+        cardNumberState: PCIFieldState,
+        expirationDateState: PCIFieldState,
+        securityCodeState: PCIFieldState,
+    ) {
+        viewModelScope.launch {
+            if (viewState.value.cardNumberState.length != viewState.value.cardNumberState.maxLength) {
+                _viewState.value = _viewState.value.copy(cardNumberState = _viewState.value.cardNumberState.copy(isError = true))
+                return@launch
+            }
+            val result = coreMethods.generateCardToken(
+                cardNumberState = cardNumberState,
+                expirationDateState = expirationDateState,
+                securityCodeState = securityCodeState,
+                buyerIdentification = BuyerIdentification(
+                    name = viewState.value.identificationState.identificationNameValue,
+                    number = viewState.value.identificationState.identificationValue,
+                    type = viewState.value.identificationState.selectedIdentification?.name
+                )
+            )
+
+            when (result) {
+                is Result.Success -> {
+                    LogHelper.log(
+                        message = "generateToken - Success",
+                        type = LogType.Network,
+                        response = gson.toJson(result.data),
+                    )
+                    onDialogStateChanged(PaymentScreenDialogState.CardToken(token = result.data.token))
+                }
+
+                is Result.Error -> {
+                    when (result.error) {
+                        is ResultError.Request -> {
+                            LogHelper.log(
+                                message = "generateToken - Error",
+                                type = LogType.Network,
+                                response = (result.error as ResultError.Request).message,
+                            )
+                            _viewState.value = _viewState.value.copy(
+                                dialogState = PaymentScreenDialogState.Error(
+                                    title = "Generate Token Error",
+                                    description = (result.error as ResultError.Request).message,
+                                ),
+                            )
+                        }
+
+                        is ResultError.Validation -> {
+                            LogHelper.log(
+                                message = "generateToken - Error",
+                                type = LogType.Network,
+                                response = (result.error as ResultError.Validation).message,
+                            )
+                            _viewState.value = _viewState.value.copy(
+                                dialogState = PaymentScreenDialogState.Error(
+                                    title = "Generate Token Error",
+                                    description = (result.error as ResultError.Validation).message,
+                                ),
+                            )
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    fun getInstallment(
+        bin: String,
+        amount: BigDecimal
+    ) {
+        viewModelScope.launch {
+            val result = coreMethods.getInstallments(
+                bin = bin,
+                amount = amount,
+            )
+
+            when (result) {
+                is Result.Success -> {
+                    LogHelper.log(
+                        message = "getInstallment - Success",
+                        type = LogType.Network,
+                        response = gson.toJson(result.data),
+                    )
+                    _viewState.value = _viewState.value.copy(
+                        installmentsState = _viewState.value.installmentsState.copy(
+                            showList = true,
+                            installments = result.data.getOrNull(0)?.payerCost?.toInstallmentModel()
+                                .orEmpty(),
+                        )
+                    )
+                }
+
+                is Result.Error -> {
+                    when (result.error) {
+                        is ResultError.Request -> {
+                            LogHelper.log(
+                                message = "getInstallment - Error",
+                                type = LogType.Network,
+                                response = (result.error as ResultError.Request).message,
+                            )
+                        }
+
+                        is ResultError.Validation -> {
+                            LogHelper.log(
+                                message = "getInstallment - Error",
+                                type = LogType.Network,
+                                response = (result.error as ResultError.Validation).message,
+                            )
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    fun getIdentificationTypes() {
+        viewModelScope.launch {
+            val result = coreMethods.getIdentificationTypes()
+            when (result) {
+                is Result.Success -> {
+                    LogHelper.log(
+                        message = "getIdentificationTypes - Success",
+                        type = LogType.Network,
+                        response = gson.toJson(result.data),
+                    )
+                    _viewState.value = _viewState.value.copy(
+                        identificationState = _viewState.value.identificationState.copy(
+                            identificationList = result.data,
+                            selectedIdentification = result.data.firstOrNull(),
+                        )
+                    )
+                }
+
+                is Result.Error -> {
+                    when (result.error) {
+                        is ResultError.Request -> {
+                            LogHelper.log(
+                                message = "getIdentificationTypes - Error",
+                                type = LogType.Network,
+                                response = (result.error as ResultError.Request).message,
+                            )
+                        }
+
+                        is ResultError.Validation -> {
+                            LogHelper.log(
+                                message = "getIdentificationTypes - Error",
+                                type = LogType.Network,
+                                response = (result.error as ResultError.Validation).message,
+                            )
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    fun getCardIssuers(bin: Int, paymentMethodId: String) {
+        viewModelScope.launch {
+            val result = coreMethods.getCardIssuers(bin, paymentMethodId)
+
+            when (result) {
+                is Result.Success -> {
+                    LogHelper.log(
+                        message = "getCardIssuers - Success",
+                        type = LogType.Network,
+                        response = gson.toJson(result.data),
+                    )
+                    _viewState.value = _viewState.value.copy(
+                        cardIssuers = result.data,
+                        cardNumberState = _viewState.value.cardNumberState.copy(image = result.data[0].thumbnail)
+                    )
+                }
+
+                is Result.Error -> {
+                    when (result.error) {
+                        is ResultError.Request -> {
+                            LogHelper.log(
+                                message = "getCardIssuers - Error",
+                                type = LogType.Network,
+                                response = (result.error as ResultError.Request).message,
+                            )
+                        }
+
+                        is ResultError.Validation -> {
+                            LogHelper.log(
+                                message = "getPaymentMethods - Error",
+                                type = LogType.Network,
+                                response = (result.error as ResultError.Validation).message,
+                            )
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    fun getPaymentMethods(bin: String) {
+        viewModelScope.launch {
+            val result = coreMethods.getPaymentMethods(bin = bin)
+
+            when (result) {
+                is Result.Success -> {
+                    LogHelper.log(
+                        message = "getPaymentMethods - Success",
+                        type = LogType.Network,
+                        response = gson.toJson(result.data),
+                    )
+                    _viewState.value = _viewState.value.copy(
+                        secureCodeState = _viewState.value.secureCodeState.copy(
+                            secureCodeLength = result.data[0].card?.securityCode?.length ?: 3
+                        )
+
+                    )
+                    getCardIssuers(
+                        bin = bin.toInt(),
+                        paymentMethodId = result.data[0].id!!
+                    )
+                    updateCardMaskState(result.data.firstOrNull()?.card?.length?.max ?: DEFAULT_MAX_CARD_LENGTH)
+                }
+
+                is Result.Error -> {
+                    when (result.error) {
+                        is ResultError.Request -> {
+                            LogHelper.log(
+                                message = "getPaymentMethods - Error",
+                                type = LogType.Network,
+                                response = (result.error as ResultError.Request).message,
+                            )
+                        }
+
+                        is ResultError.Validation -> {
+                            LogHelper.log(
+                                message = "getPaymentMethods - Error",
+                                type = LogType.Network,
+                                response = (result.error as ResultError.Validation).message,
+                            )
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    fun onExpirationDateEvent(event: ExpirationDateTextFieldEvent) {
+        LogHelper.log(
+            message = "onExpirationDateEvent - $event",
+            type = LogType.Function,
+            response = event.toString(),
+        )
+        when (event) {
+            is ExpirationDateTextFieldEvent.OnInputFilled -> {
+                _viewState.value = _viewState.value.copy(
+                    expirationDateState = _viewState.value.expirationDateState.copy(
+                        filled = event.isFilled
+                    ),
+                )
+            }
+
+            is ExpirationDateTextFieldEvent.IsValid -> {
+                _viewState.value = _viewState.value.copy(
+                    expirationDateState = _viewState.value.expirationDateState.copy(
+                        valid = event.isValid
+                    )
+                )
+            }
+
+            is ExpirationDateTextFieldEvent.OnFocusChanged -> {
+                _viewState.value = _viewState.value.copy(
+                    expirationDateState = _viewState.value.expirationDateState.copy(
+                        isFocused = event.isFocused
+                    )
+                )
+            }
+
+            is ExpirationDateTextFieldEvent.OnLengthChanged -> {
+                _viewState.value = _viewState.value.copy(
+                    expirationDateState = _viewState.value.expirationDateState.copy(
+                        length = event.length
+                    )
+                )
+            }
+        }
+    }
+
+    fun onSecurityCodeEvent(event: SecurityCodeTextFieldEvent) {
+        LogHelper.log(
+            message = "onSecurityCodeEvent - $event",
+            type = LogType.Function,
+            response = event.toString(),
+        )
+        when (event) {
+            is SecurityCodeTextFieldEvent.OnFocusChanged -> {
+                _viewState.value = _viewState.value.copy(
+                    secureCodeState = _viewState.value.secureCodeState.copy(
+                        isFocused = event.isFocused
+                    )
+                )
+            }
+
+            is SecurityCodeTextFieldEvent.OnLengthChanged -> {
+                _viewState.value = _viewState.value.copy(
+                    secureCodeState = _viewState.value.secureCodeState.copy(
+                        length = event.length
+                    )
+                )
+            }
+
+            is SecurityCodeTextFieldEvent.OnInputFilled -> {
+                _viewState.value = _viewState.value.copy(
+                    secureCodeState = _viewState.value.secureCodeState.copy(
+                        filled = event.isFilled
+                    )
+                )
+            }
+        }
+    }
+
+    fun onCardNumberEvent(event: CardNumberTextFieldEvent) {
+        LogHelper.log(
+            message = "onCardNumberEvent - $event",
+            type = LogType.Function,
+            response = event.toString(),
+        )
+        when (event) {
+            is CardNumberTextFieldEvent.OnFocusChanged -> {
+                _viewState.value = _viewState.value.copy(
+                    cardNumberState = _viewState.value.cardNumberState.copy(
+                        isFocused = event.isFocused
+                    )
+                )
+            }
+
+            is CardNumberTextFieldEvent.OnLengthChanged -> {
+                _viewState.value = _viewState.value.copy(
+                    cardNumberState = _viewState.value.cardNumberState.copy(
+                        length = event.length,
+                        isError = false,
+                    )
+                )
+            }
+
+            is CardNumberTextFieldEvent.OnLastFourDigitsFilled -> {
+                _viewState.value = _viewState.value.copy(
+                    cardNumberState = _viewState.value.cardNumberState.copy(
+                        lastFourDigits = event.lastFourDigits
+                    )
+                )
+            }
+
+            is CardNumberTextFieldEvent.IsValid -> {
+                _viewState.value = _viewState.value.copy(
+                    cardNumberState = _viewState.value.cardNumberState.copy(
+                        isValid = event.isValid
+                    )
+                )
+            }
+
+            is CardNumberTextFieldEvent.OnBinChanged -> {
+                if ((event.cardBin?.length ?: 0) < CARD_NUMBER_BIN_LENGTH) {
+                    _viewState.value =
+                        _viewState.value.copy(
+                            cardNumberState = _viewState.value.cardNumberState.copy(image = null),
+                            installmentsState = _viewState.value.installmentsState.copy(showList = false)
+                        )
+                    updateCardMaskState(DEFAULT_MAX_CARD_LENGTH)
+                } else {
+                    getInstallment(
+                        bin = event.cardBin.orEmpty(),
+                        amount = 1000.0.toBigDecimal(),
+                    )
+                    getPaymentMethods(bin = event.cardBin.orEmpty())
+                }
+
+                _viewState.value = _viewState.value.copy(
+                    cardNumberState = _viewState.value.cardNumberState.copy(
+                        cardBin = event.cardBin
+                    )
+                )
+
+                getInstallment(
+                    bin = event.cardBin.orEmpty(),
+                    amount = 1000.0.toBigDecimal(),
+                )
+            }
+        }
+    }
+
+    fun onInstallmentSelected(value: Installment) {
+        _viewState.value = _viewState.value.copy(
+            installmentsState = _viewState.value.installmentsState.copy(
+                selectedInstallment = value
+            )
+        )
+    }
+
+    fun onIdentificationTypeValueChanged(value: String) {
+        _viewState.value = _viewState.value.copy(
+            identificationState = _viewState.value.identificationState.copy(
+                identificationValue = value
+            )
+        )
+    }
+
+    fun onIdentificationTypeChanged(identificationType: IdentificationType) {
+        _viewState.value = _viewState.value.copy(
+            identificationState = _viewState.value.identificationState.copy(
+                selectedIdentification = identificationType,
+            )
+        )
+    }
+
+    fun onCardHolderNameChanged(value: String) {
+        _viewState.value = _viewState.value.copy(
+            identificationState = _viewState.value.identificationState.copy(
+                identificationNameValue = value
+            )
+        )
+    }
+
+    fun onDialogStateChanged(dialogState: PaymentScreenDialogState) {
+        _viewState.value = _viewState.value.copy(dialogState = dialogState)
+    }
+
+    private fun updateCardMaskState(cardLength: Int) {
+        _viewState.value = _viewState.value.copy(
+            cardNumberState = _viewState.value.cardNumberState.copy(
+                maxLength = cardLength,
+                mask = when (cardLength) {
+                    8 -> "#### ####"
+                    9 -> "#### #####"
+                    10 -> "#### ######"
+                    11 -> "#### #### ###"
+                    12 -> "#### #### ####"
+                    13 -> "#### ###### ###"
+                    14 -> "#### ###### ####"
+                    15 -> "#### ###### #####"
+                    16 -> "#### #### #### ####"
+                    17 -> "#### #### #### #####"
+                    19 -> "#### #### #### #### ###"
+                    else -> "#### #### #### ####"
+                }
+            )
+        )
+    }
+}
