--- example/src/main/java/com/mercadopago/sdk/android/example/presentation/data/Installment.kt
@@ -1,5 +1,0 @@
-package com.mercadopago.sdk.android.example.presentation.data
-
-internal data class Installment(
-    val value: String
-)
