+++ example/src/main/java/com/mercadopago/sdk/android/example/presentation/logs/LogsViewState.kt
@@ -1,0 +1,9 @@
+package com.mercadopago.sdk.android.example.presentation.logs
+
+import com.mercadopago.sdk.android.example.domain.model.LogEntry
+import com.mercadopago.sdk.android.example.domain.model.LogType
+
+internal data class LogsViewState(
+    val logsList: List<LogEntry> = emptyList(),
+    val currentFilter: LogType = LogType.All,
+)
