+++ example/src/main/java/com/mercadopago/sdk/android/example/presentation/theme/Theme.kt
@@ -1,0 +1,63 @@
+package com.mercadopago.sdk.android.example.presentation.theme
+
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Typography
+import androidx.compose.material3.darkColorScheme
+import androidx.compose.material3.lightColorScheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.sp
+
+private val DarkColorScheme = darkColorScheme(
+    primary = Color(0xFFFFD700),
+    onPrimary = Color(0xFF000000),
+    secondary = Color(0xFF03DAC6),
+    onSecondary = Color(0xFF000000),
+    background = Color(0xFF121212),
+    surface = Color(0xFF1E1E1E),
+    onBackground = Color(0xFFE0E0E0),
+    onSurface = Color(0xFFE0E0E0),
+    onSurfaceVariant = Color(0xFFA0A0A0),
+)
+
+private val LightColorScheme = lightColorScheme(
+    primary = Color(0xFF3483FA),
+    onPrimary = Color(0xFFFFFFFF),
+    background = Color(0xFFFFFFFF),
+    onBackground = Color(0xFF1A1A1A),
+    onSurfaceVariant = Color(0xFF737373),
+    secondary = Color(0xFFFFE600),
+    tertiary = Color(0xFFD0BCFF),
+    outline = Color(0xFFBFBFBF),
+)
+
+private val Typography = Typography(
+    bodyLarge = TextStyle(
+        fontFamily = FontFamily.Default,
+        fontWeight = FontWeight.Normal,
+        fontSize = 16.sp,
+        lineHeight = 24.sp,
+        letterSpacing = 0.5.sp
+    )
+)
+
+@Composable
+internal fun MercadoPagoSampleTheme(
+    darkTheme: Boolean = isSystemInDarkTheme(),
+    content: @Composable () -> Unit
+) {
+    val colorScheme = when {
+        darkTheme -> DarkColorScheme
+        else -> LightColorScheme
+    }
+
+    MaterialTheme(
+        colorScheme = colorScheme,
+        typography = Typography,
+        content = content
+    )
+}
