+++ sdk-android/src/main/java/com/mercadopago/sdk/android/di/datasource/DataSourceModule.kt
@@ -1,9 +1,9 @@
 package com.mercadopago.sdk.android.di.datasource
 
+import android.content.Context
 import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.PreferenceDataStoreFactory
 import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.preferencesDataStoreFile
+import androidx.datastore.preferences.preferencesDataStore
 import com.mercadopago.sdk.android.data.local.datasource.SdkInitializationLocalDataSource
 import com.mercadopago.sdk.android.data.local.datasource.SdkInitializationLocalDataSourceImpl
 import com.mercadopago.sdk.android.data.remote.datasource.SdkInitializationRemoteDataSource
@@ -16,6 +16,10 @@ import java.security.MessageDigest
 
 private const val MESSAGE_DIGEST_SHA_256 = "MESSAGE_DIGEST_SHA_256"
 private const val SDK_INITIALIZER_DEFAULT_DATA_STORE = "SDK_INITIALIZER_DEFAULT_DATA_STORE"
+
+private val Context.sdkDataStore: DataStore<Preferences> by preferencesDataStore(
+    name = SDK_INITIALIZER_DEFAULT_DATA_STORE
+)
 
 internal fun provideDataSourceModule(): Module = module {
     factory<SdkInitializationLocalDataSource> {
@@ -33,8 +37,6 @@ internal fun provideDataSourceModule(): Module = module {
         MessageDigest.getInstance("SHA-256")
     }
     single<DataStore<Preferences>>(named(SDK_INITIALIZER_DEFAULT_DATA_STORE)) {
-        PreferenceDataStoreFactory.create {
-            androidApplication().preferencesDataStoreFile(SDK_INITIALIZER_DEFAULT_DATA_STORE)
-        }
+        androidApplication().sdkDataStore
     }
 }
